Ensure deferred cleanup when display is off
Using DRM video playback app and wi-fi direct, allocation of
AutoBackendTexture objects sometimes continue for some time after the
display is off. Clean-up of these is deferred until postRender() is
called, but only invoked if the display is on. This eventually leads
to depletion of secure memory resources and a fatal assert.

This ensures cleanupPostRender() is called regardless of display
state.

Test: Play Movies with screen mirroring, repeatedly sleep/wake device
Change-Id: Ib0d530582e124984ef70f91f6e398f62c2312f4c
Signed-off-by: rajkale99 <kaleraj.rk@gmail.com>